add task 0 fix

+++ 0-bubble_sort.c
@@ -8,7 +8,7 @@
 */
 void bubble_sort(int *array, size_t size)
 {
-	int i, j, temp, swap;
+	size_t i, j, temp, swap;
 
 	swap = 0;
